fix: dev container config file updates
The devcontainer config file was a bit outdated, and a lot of
configuration/scripts have changed since then. I was unable to
start a new dev container inside my vscode, so I forked over
the latest scripts and updated the configurations so that it
runs with the
[0.309.0](https://github.com/microsoft/vscode-remote-release/issues/9060#issuecomment-1747811300)
version of the devcontainer vscode extension.

Feel free to crosscheck the config files with the latest release 0.245.2
of the [here](https://github.com/microsoft/vscode-dev-containers/tree/v0.245.2/containers/codespaces-linux/.devcontainer).

no related PRs